feat: Setup for testing
Install Jest. Install Babel
and its dependency according
to the docs.

Configure Babel to target the
current version of Node in a
babel.config.js file.

Setup scripts to in package.json
to run Jest and another
to watch files and test on save.

Create gitignore and add
node_modules to it.

Create index.html and index.js
files to be able to serve
files on the browser.

Populate the index.html
with boilerplate and
in script tags
set index.js to be
handled as a module.