Initialize CaesarCipher
Create a caesar-cypher
file and in it,
initialize
and export a
CaesarCipher class
with shift, string and
plainTextAlphabets
instance variables.

Initialize a shiftLetters
method that takes in an
offset and returns
cipher text alphabets
from that offset.

It has a for loop that for
each letter of the plain
text alphabet, it adds
its index with offset
and takes the modulo
of the result with the
length of the plain
text alphabets. It
saves the result of
this in a shiftedIndex
and use on the
plain text alphabet to
obtain the
cipherTextLetter
equivalent.

Concatenate all cipher
text alphabets and
return them as a string.