Implement encrypt
Initialize an encrypt
method that takes in
a shift and a string
and returns an encrypted
string from the
shift value.

First get the
cipher text alphabets
by calling shiftLetters
on the shift arg.

In a for loop,
Get each letter in the
string's `index` in
the plain text alphabets.

Find its equivalent in
the cipher text alphabet
by using its `index`
on the cipherTextAlphabets
string.

Concatenate the encrypted
string and return it.