Update CaesarCipher to preserve case
In the test file, write a test
to check if the encryption
preserves case.

In the caesar-cipher file,
update CaesarCipher to
preserve case.

First rename the alphabet
instance variable to
a suitable name. Update
its references accordingly.
Add a new instance
variable which contains
uppercases of the plain
alphabet.

Update shiftLetters()
to have lowercase and
uppercase variables
(versions of
the cipher alphabet}.
Use shiftIndex to get
the upper and lower
cases and return them
as an object.

In encrypt, rename
cypherTextAlphabets
to a suitable name
and update its
references
accordingly.

Instead of an empty
string,
initialize an array
that will hold the
letters of the
encryped test.

In the for loop,
instead of fetching
the index differently
and concatenating
the letters differently,
find where the letter
is found.

If it is not found in
the lower case alphabets
(plain),
check in upper case
(plain)
alphabets, get the
index of where this
letter is found.
Set `letter` to be
the letter that is at
the this respective
index in the uppercase
cipher letters.
Else(it was found in
plain lower case alphabets)
get the respective index and
Set `letter` to be
the letter that is at
the this respective
index in the lower
cipher letters.

This preserves the case.