Update to preserve other characters
In the test file, write a test
to check if the encryption
preserves punctuation and other
non-alphabetic characters.

In the caesar-cypher file
in the encrypt method, define
a letter variable and update
the ternary operator to a
conditional statement.
The coditional statement
checks for the presence of
the letter in the
upper and lower case plain
alphabets gets its index
and sets letter to
the equivalent index in
either the upper or lower
case cipher alphabets. Else
(it means its a character
that isnt found in either
the upper or lower case
cipher alphabets), set letter
to be the character which
wasn't found in the upper
and lower case
cipher alphabets.

This preserves space,
punctuation and
other non-alphabetic
characters.