Implement AnalyzeArray
Create analyze-array file
and its test file.

In the test file, write a
test for an analyze array
class that returns an
object with certain
properties.

In analyzeArray, initialize
an AnalyzeArray class.
Implement getAverage, getMax,
getMin and getLength methods.
These methods take in an
array and returns the average,
maximum value, minimum value
and length of the array
respectively.

Implement an analyze method
that takes in an array,
and returns an object with
properties.
Set the values of these
properties to the result
of calling the
getAverage, getMax,
getMin and getLength
respectively on
the array which is passed
into the analyze method.

This passes the test.